Add a custom LDAP strategy to permit better tweaks of behaviour.
For example, the default OmniAuth-LDAP strategy considers 'cn' for the
'name' field, but our AD server returns the login account name for
the 'cn' (i.e. 'jdoe') rather than the 'displayName' (i.e. 'John Doe').
The mapping behaviour is held in a class variable `@@config`, but this
is only used in the `map_user` function, so the cleanest option is to
subclass the strategy and override that function. We can ensure that
the provider still uses `ldap` as the name by explicitly setting it
when building the authentication provider (otherwise it will attempt
to convert the class name).