Move to the maintained Gitlab fork of OmniAuth-LDAP.
The original 'omniauth-ldap' has had very few commits since the v1.0.5
release in 2016 used by this plugin, and on RubyGems it is clear that
the Gitlab fork is better maintained with a number of releases since,
up to the current v2.2.0 release in 2022. This is also reflected in the
total downloads; the original gem has had 719k total downloads (of which
449k for v1.0.5) whilst the Gitlab fork has 30m total downloads (of
which 16m for v2.2.0).

NB: The Gitlab fork replaces :method with :encryption, and the values
are now 'plain', 'simple_tls' (ssl) and 'start_tls' (tls) instead of
'plain', 'ssl', 'tls'. However the :method option still exists within
the adaptor (albeit deprecated), and it will convert the values
accordingly. The rest of the code does not appear to have any breaking
changes, but has added support for better error handling and also
better encryption settings.

As part of this update we also update the supporting LDAP and NTLM
gems to their latest supported versions for this gem, to keep up with
any potential security issues.